+++ AsyncDisplayKit/Layout/ASLayoutable.h
@@ -20,53 +20,6 @@
  */
 @protocol ASLayoutable <NSObject>
 
-/** 
- * @abstract Additional space to place before this object in the stacking direction.
- * Used when attached to a stack layout.
- */
-@property (nonatomic, readwrite) CGFloat spacingBefore;
-
-/**
- * @abstract Additional space to place after this object in the stacking direction.
- * Used when attached to a stack layout.
- */
-@property (nonatomic, readwrite) CGFloat spacingAfter;
-
-/**
- * @abstract If the sum of childrens' stack dimensions is less than the minimum size, should this object grow?
- * Used when attached to a stack layout.
- */
-@property (nonatomic, readwrite) BOOL flexGrow;
-
-/**
- * @abstract If the sum of childrens' stack dimensions is greater than the maximum size, should this object shrink?
- * Used when attached to a stack layout.
- */
-@property (nonatomic, readwrite) BOOL flexShrink;
-
-/**
- * @abstract Specifies the initial size in the stack dimension for this object.
- * Default to ASRelativeDimensionUnconstrained.
- * Used when attached to a stack layout.
- */
-@property (nonatomic, readwrite) ASRelativeDimension flexBasis;
-
-/**
- * @abstract Orientation of the object along cross axis, overriding alignItems
- * Used when attached to a stack layout.
- */
-@property (nonatomic, readwrite) ASStackLayoutAlignSelf alignSelf;
-
-/**
- *  @abstract Used for baseline alignment. The distance from the top of the object to its baseline.
- */
-@property (nonatomic, readwrite) CGFloat ascender;
-
-/**
- *  @abstract Used for baseline alignment. The distance from the baseline of the object to its bottom.
- */
-@property (nonatomic, readwrite) CGFloat descender;
-
 /**
  * @abstract Calculate a layout based on given size range.
  *
