+++ AsyncDisplayKit/Private/ASStackPositionedLayout.h
@@ -17,8 +17,7 @@
 struct ASStackPositionedLayout {
   const std::vector<ASLayout *> sublayouts;
   const CGFloat crossSize;
-  const CGFloat ascender;
-  const CGFloat descender;
+  const CGFloat distanceToBaseline;
 
   /** Given an unpositioned layout, computes the positions each child should be placed at. */
   static ASStackPositionedLayout compute(const ASStackUnpositionedLayout &unpositionedLayout,
